blob_api/requirements.txt
asgiref==3.3.1
azure-core==1.11.0
azure-storage-blob==12.7.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
ftfy==5.9
idna==2.10
isodate==0.6.0
jdcal==1.4.1
msrest==0.6.21
numpy==1.20.1
oauthlib==3.1.0
openpyxl==3.0.6
pandas==1.2.2
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
wcwidth==0.2.5
wincertstore==0.2
xlrd==2.0.1
